Moved `Parser` class to separate file, outside of the `RuleBasedParser`. Moved `cruise.umple.parser` code to `UmpleParser/` directory. Updated build-scripts. Added readme.
