+++ cruise.umple/src/Master.ump
@@ -24,10 +24,7 @@ use Documenter.ump;
 use Generator.ump;
 use Json.ump;
 use Parser.ump;
-use ParsingRules.ump;
-use GrammarParsing.ump;
 use StateMachine.ump;
-use TextParser.ump;
 use Umlet.ump;
 use Umple.ump;
 use Trace.ump;
